pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dfeqa"
authors = [
  { name="mrheducation", email="martin1.harris@education.gov.uk" },
]
description = "helpers for quality checking data"
readme = "README.md"
dynamic = ["version"]
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "pandas",
  "sqlalchemy",
  "seaborn",
  "regex",
  "jinja2",
  "python-dotenv",
]

[project.optional-dependencies]
user = [
  "quarto",
  "pyyaml",
  "nbformat",
  "nbclient",
  "ipykernel",
]
databricks = [
  "dfeqa[user]",
  "databricks-sql-connector",
  "databricks-sqlalchemy"
]
odbc = [
  "dfeqa[user]",
  "pyodbc"
]

[project.urls]
Documentation = "https://github.com/quhack/dfeqa/wiki"
Repository = "https://github.com/quhack/dfeqa"
Issues = "https://github.com/quhack/dfeqa/issues"

[project.scripts]
dfeqa = "dfeqa.cmdline:execute"

[tool.pytest.ini_options]
markers = [
  "slow",
]

addopts = [
    "--import-mode=importlib",
    "--cov=src",
#    "--cov-append",
    "--cov-report=html",
    "--cov-report=term-missing",
#    "--cov-fail-under=95",
]
testpaths = [
  "tests"
]

[tool.coverage.run]
  source = ["src"]

[tool.hatch.build]
  sources = ["src"]

[tool.hatch.build.targets.sdist]
only-include = [
  "src/dfeqa",
  "tests",
]

[tool.hatch.build.targets.wheel]
  sources = ["src"]

[tool.hatch.version]
path = "src/dfeqa/__about__.py"
